Fixing tested maximum tested level in test_node.
Also adding assertion in node::set_level to check the given level.

For unsigned short type, 4 levels was tested in test_node.cpp when this type
can only handle 3 levels. It was triggering an out-of-index exception on
Visual Studio with clang compiler leading to a test failure.

Don't know why other compilers on Linux platform don't do the same...